+++ data/src/main/java/daily/dayo/data/datasource/remote/heart/HeartRequest.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class CreateHeartRequest(
     @SerializedName("postId")
-    val postId: Int
+    val postId: Long
 )
