+++ domain/src/main/java/daily/dayo/domain/usecase/bookmark/RequestDeleteBookmarkPostUseCase.kt
@@ -6,6 +6,6 @@ import javax.inject.Inject
 class RequestDeleteBookmarkPostUseCase @Inject constructor(
     private val bookmarkRepository: BookmarkRepository
 ) {
-    suspend operator fun invoke(postId: Int) =
+    suspend operator fun invoke(postId: Long) =
         bookmarkRepository.requestDeleteBookmarkPost(postId)
 }
